+++ backend/src/main/java/com/cronoattendance/repositories/DepartamentosRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.cronoattendance.entities.Departamentos;
 
 @Repository
-public interface DepartamentosRepository extends JpaRepository<Departamentos, Long> {
+public interface DepartamentosRepository extends JpaRepository<Departamentos, Integer> {
 }
